# openapi/account-api.yaml
openapi: 3.1.0
info:
  title: Account Center Management API
  version: 1.0.0
  description: >-
    Account 管理端同源 Session API。浏览器不得持有应用 Secret、Session ID 或长期 Token。
    管理端 message/fieldErrors 根据可选 Accept-Language 返回中文或英文；支持 zh/en 地区变体与权重，
    缺失、不支持或非法头回退 zh-CN。响应包含 Content-Language 和 Vary: Accept-Language。
    错误码、字段名、状态值及用户业务数据不随语言变化。
servers:
  - url: /
...
paths:
  /api/session:
    get:
      tags: [Session]
      operationId: getSession
      summary: 初始化或读取当前管理会话
      responses:
        '200':
          description: 当前会话；未登录时 authenticated 为 false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
    post:
      tags: [Session]
      operationId: login
      summary: 使用 Account 管理身份登录
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功，服务端轮换 Session ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    delete:
      tags: [Session]
      operationId: logout
      summary: 退出并使当前 Session 失效
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: 已退出
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/audit-events:
    get:
      tags: [Audit]
      operationId: listAuditEvents
      summary: 分页查询脱敏审计事件
      description: Admin/Auditor 只读；精确筛选；固定按 createdAt desc、id desc 排序，不返回原始 detail。
      security:
        - sessionCookie: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100000, default: 1 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: operatorId
          in: query
          schema: { type: string, maxLength: 64 }
        - name: operationType
          in: query
          schema: { type: string, maxLength: 64 }
        - name: targetType
          in: query
          schema: { type: string, maxLength: 64 }
        - name: targetId
          in: query
          schema: { type: string, maxLength: 128 }
        - name: traceId
          in: query
          schema: { type: string, maxLength: 64 }
      responses:
        '200':
          description: 脱敏事件分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/audit-events/{id}:
    get:
      tags: [Audit]
      operationId: getAuditEvent
      summary: 获取脱敏审计事件详情
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 64 }
      responses:
        '200':
          description: 只读事件，不返回原始 detail 或票据目标标识
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
      description: 由 GET /api/session 初始化并只保存在前端内存中的同步器 Token
...
  responses:
    Unauthorized:
      description: 未登录或 Session 已失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: CSRF 校验失败或当前角色没有访问能力
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: 请求字段校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    AuditEventResponse:
      type: object
      additionalProperties: false
      required: [id, operatorId, operationType, targetType, targetId, outcome, traceId, createdAt]
      properties:
        id: { type: string }
        operatorId: { type: [string, 'null'] }
        operationType: { type: string }
        targetType: { type: [string, 'null'] }
        targetId:
          type: [string, 'null']
          description: 仅 USER、APPLICATION、USER_APPLICATION 管理目标公开，其他类型隐藏。
        outcome: { type: string, enum: [success, failure, unknown] }
        traceId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    AuditEventPageResponse:
      type: object
      additionalProperties: false
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditEventResponse'
        page: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, format: int64, minimum: 0 }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [account, password]
      properties:
        account:
          type: string
          minLength: 1
          maxLength: 128
        password:
          type: string
          minLength: 1
          maxLength: 256
          writeOnly: true
    SessionResponse:
      type: object
      additionalProperties: false
      required: [authenticated, user, roles, capabilities, csrfToken]
      properties:
        authenticated:
          type: boolean
        user:
          oneOf:
            - $ref: '#/components/schemas/SessionUser'
            - type: 'null'
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [ACCOUNT_ADMIN, ACCOUNT_AUDITOR]
        capabilities:
          type: array
          uniqueItems: true
          items:
            type: string
            enum:
              - 'users:read'
              - 'users:write'
              - 'applications:read'
              - 'applications:write'
              - 'application-access:read'
              - 'application-access:write'
        csrfToken:
          type: string
          minLength: 32
    SessionUser:
      type: object
      additionalProperties: false
      required: [id, account, name]
      properties:
        id:
          type: string
        account:
          type: string
        name:
          type: string
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, traceId]
      properties:
        code:
          type: string
          examples: [AUTHENTICATION_REQUIRED]
        message:
          type: string
        traceId:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
...